refactor: use .graphql extension for auth queries and mutations
