+++ src/apollo/auth/mutations/LogOut.mutation.graphql
@@ -1,0 +1,3 @@
+mutation LogOut {
+  logOut
+}
