--- src/apollo/auth/mutations/LogOut.mutation.ts
@@ -1,7 +1,0 @@
-import { gql } from "@apollo/client";
-
-gql`
-  mutation LogOut {
-    logOut
-  }
-`;
