+++ src/apollo/auth/mutations/Login.mutation.graphql
@@ -1,0 +1,7 @@
+mutation Login($input: LoginInput!) {
+  login(input: $input) {
+    user {
+      ...UserAvatar
+    }
+  }
+}
