--- src/apollo/auth/mutations/Login.mutation.ts
@@ -1,13 +1,0 @@
-import { gql } from "@apollo/client";
-import { UserAvatarFragmentDoc } from "../../gen";
-
-gql`
-  mutation Login($input: LoginInput!) {
-    login(input: $input) {
-      user {
-        ...UserAvatar
-      }
-    }
-  }
-  ${UserAvatarFragmentDoc}
-`;
