+++ src/apollo/auth/mutations/RefreshToken.mutation.graphql
@@ -1,0 +1,3 @@
+mutation RefreshToken {
+  refreshToken
+}
