--- src/apollo/auth/mutations/RefreshToken.mutation.ts
@@ -1,23 +1,0 @@
-import { gql } from "@apollo/client";
-import client from "../../client";
-import { logOutUser } from "../../../utils/auth.utils";
-import { isRefreshingTokenVar } from "../../cache";
-import { RefreshTokenDocument } from "../../gen";
-
-export const refreshToken = async () => {
-  try {
-    isRefreshingTokenVar(true);
-    await client.mutate({ mutation: RefreshTokenDocument });
-  } catch (err) {
-    await logOutUser();
-    throw err;
-  } finally {
-    isRefreshingTokenVar(false);
-  }
-};
-
-gql`
-  mutation RefreshToken {
-    refreshToken
-  }
-`;
