+++ src/apollo/auth/mutations/SignUp.mutation.graphql
@@ -1,0 +1,7 @@
+mutation SignUp($input: SignUpInput!) {
+  signUp(input: $input) {
+    user {
+      ...UserAvatar
+    }
+  }
+}
