--- src/apollo/auth/mutations/SignUp.mutation.ts
@@ -1,13 +1,0 @@
-import { gql } from "@apollo/client";
-import { UserAvatarFragmentDoc } from "../../gen";
-
-gql`
-  mutation SignUp($input: SignUpInput!) {
-    signUp(input: $input) {
-      user {
-        ...UserAvatar
-      }
-    }
-  }
-  ${UserAvatarFragmentDoc}
-`;
