+++ src/apollo/auth/queries/AuthCheck.query.graphql
@@ -1,0 +1,3 @@
+query AuthCheck {
+  authCheck
+}
