--- src/apollo/auth/queries/AuthCheck.query.ts
@@ -1,7 +1,0 @@
-import { gql } from "@apollo/client";
-
-gql`
-  query AuthCheck {
-    authCheck
-  }
-`;
