+++ src/utils/common.utils.ts
@@ -6,7 +6,7 @@ import { t } from "i18next";
 import Router from "next/router";
 import React, { isValidElement, ReactNode } from "react";
 import { animateScroll } from "react-scroll";
-import { refreshToken } from "../apollo/auth/mutations/RefreshToken.mutation";
+import { refreshToken } from "../apollo/auth/links/refresh-token.link";
 import { isRefreshingTokenVar, toastVar } from "../apollo/cache";
 import {
   API_ROOT,
